--- dubbo-config/dubbo-config-spring/src/main/java/org/apache/dubbo/config/spring/util/BeanRegistrar.java
@@ -1,64 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.dubbo.config.spring.util;
-
-import org.springframework.beans.factory.config.BeanDefinition;
-import org.springframework.beans.factory.support.BeanDefinitionRegistry;
-import org.springframework.beans.factory.support.RootBeanDefinition;
-import org.springframework.core.AliasRegistry;
-
-import static org.springframework.util.ObjectUtils.containsElement;
-import static org.springframework.util.StringUtils.hasText;
-
-/**
- * Bean Registrar
- *
- * @since 2.5.7
- */
-public class BeanRegistrar {
-
-    /**
-     * Register Infrastructure Bean
-     *
-     * @param beanDefinitionRegistry {@link BeanDefinitionRegistry}
-     * @param beanType               the type of bean
-     * @param beanName               the name of bean
-     */
-    public static void registerInfrastructureBean(BeanDefinitionRegistry beanDefinitionRegistry,
-                                                  String beanName,
-                                                  Class<?> beanType) {
-
-        if (!beanDefinitionRegistry.containsBeanDefinition(beanName)) {
-            RootBeanDefinition beanDefinition = new RootBeanDefinition(beanType);
-            beanDefinition.setRole(BeanDefinition.ROLE_INFRASTRUCTURE);
-            beanDefinitionRegistry.registerBeanDefinition(beanName, beanDefinition);
-        }
-
-    }
-
-    /**
-     * Detect the alias is present or not in the given bean name from {@link AliasRegistry}
-     *
-     * @param registry {@link AliasRegistry}
-     * @param beanName the bean name
-     * @param alias    alias to test
-     * @return if present, return <code>true</code>, or <code>false</code>
-     */
-    public static boolean hasAlias(AliasRegistry registry, String beanName, String alias) {
-        return hasText(beanName) && hasText(alias) && containsElement(registry.getAliases(beanName), alias);
-    }
-}
