--- dubbo-config/dubbo-config-spring/src/main/java/org/apache/dubbo/config/spring/util/PropertySourcesUtils.java
@@ -1,115 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.dubbo.config.spring.util;
-
-import org.springframework.core.env.EnumerablePropertySource;
-import org.springframework.core.env.MutablePropertySources;
-import org.springframework.core.env.PropertyResolver;
-import org.springframework.core.env.PropertySource;
-import org.springframework.core.env.PropertySources;
-import org.springframework.core.env.PropertySourcesPropertyResolver;
-
-import java.util.Iterator;
-import java.util.LinkedHashMap;
-import java.util.Map;
-import java.util.Properties;
-
-import static java.util.Collections.unmodifiableMap;
-
-
-/**
- * {@link PropertySources} Utilities
- * <p>
- *
- * @since 2.6.6
- */
-public abstract class PropertySourcesUtils {
-
-    /**
-     * Get prefixed {@link Properties}
-     *
-     * @param propertySources {@link PropertySource} Iterable
-     * @param prefix          the prefix of property name
-     * @return Map
-     * @see Properties
-     */
-    public static Map<String, Object> getPrefixedProperties(Iterable<PropertySource<?>> propertySources, String prefix) {
-
-        MutablePropertySources mutablePropertySources = new MutablePropertySources();
-
-        for (PropertySource<?> source : propertySources) {
-            mutablePropertySources.addLast(source);
-        }
-
-        return getPrefixedProperties(mutablePropertySources, prefix);
-
-    }
-
-    /**
-     * Get prefixed {@link Properties}
-     *
-     * @param propertySources {@link PropertySources}
-     * @param prefix          the prefix of property name
-     * @return Map
-     * @see Properties
-     */
-    public static Map<String, Object> getPrefixedProperties(PropertySources propertySources, String prefix) {
-
-        PropertyResolver propertyResolver = new PropertySourcesPropertyResolver(propertySources);
-
-        Map<String, Object> prefixedProperties = new LinkedHashMap<>();
-
-        String normalizedPrefix = buildPrefix(prefix);
-
-        Iterator<PropertySource<?>> iterator = propertySources.iterator();
-
-        while (iterator.hasNext()) {
-            PropertySource<?> source = iterator.next();
-            if (source instanceof EnumerablePropertySource) {
-                for (String name : ((EnumerablePropertySource<?>) source).getPropertyNames()) {
-                    if (!prefixedProperties.containsKey(name) && name.startsWith(normalizedPrefix)) {
-                        String subName = name.substring(normalizedPrefix.length());
-                        if (!prefixedProperties.containsKey(subName)) { // take first one
-                            Object value = source.getProperty(name);
-                            if (value instanceof String) {
-                                // Resolve placeholder
-                                value = propertyResolver.resolvePlaceholders((String) value);
-                            }
-                            prefixedProperties.put(subName, value);
-                        }
-                    }
-                }
-            }
-        }
-
-        return unmodifiableMap(prefixedProperties);
-    }
-
-    /**
-     * Build the prefix
-     *
-     * @param prefix the prefix
-     * @return the prefix
-     */
-    public static String buildPrefix(String prefix) {
-        if (prefix.endsWith(".")) {
-            return prefix;
-        } else {
-            return prefix + ".";
-        }
-    }
-}
